requirements.txt
mlx-lm
mlx-data
numpy
mteb
elasticsearch
wandb
